@@ -1,48 +1,39 @@
 {
-  "name": "@lk77/vue3-color",
-  "version": "3.0.6",
-  "description": "Color of Vue 3 Components",
-  "keywords": [
-    "color",
-    "vuejs",
-    "vuejs3",
-    "vue3"
-  ],
-  "main": "dist/vue3-color.common.js",
-  "unpkg": "dist/vue3-color.umd.min.js",
-  "jsdelivr": "dist/vue3-color.umd.min.js",
+  "name": "v-color-picker",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "dev": "npm run serve",
-    "prod": "npm run build && npm run gh:build",
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build --target lib --name vue3-color src/build.js",
-    "gh:build": "vue-cli-service build --dest='dist/vue3-color'",
-    "lint": "vue-cli-service lint",
-    "patch": "npm version patch && npm run build && npm publish && git push --follow-tags"
+    "dev": "vite",
+    "build": "vue-tsc -b && vite build",
+    "preview": "vite preview"
   },
-  "homepage": "https://lk77.github.io/vue3-color/",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/lk77/vue3-color"
+  "dependencies": {
+    "@vueuse/core": "^11.2.0",
+    "lodash-es": "^4.17.21",
+    "tinycolor2": "^1.6.0",
+    "vue": "^3.5.12"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "@vue/eslint-config-standard": "^6.1.0",
-    "clamp": "^1.0.1",
-    "core-js": "^3.21.1",
-    "eslint": "^7.32.0",
-    "eslint-plugin-import": "^2.25.3",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-vue": "^8.0.3",
-    "lodash": "^4.17.21",
-    "material-colors": "^1.2.6",
-    "tinycolor2": "^1.4.2",
-    "vue": "^3.2.31"
-  },
-  "license": "MIT"
+    "@eslint/js": "^9.14.0",
+    "@types/lodash-es": "^4.17.12",
+    "@types/tinycolor2": "^1.4.6",
+    "@vitejs/plugin-vue": "^5.1.4",
+    "autoprefixer": "^10.4.20",
+    "class-variance-authority": "^0.7.0",
+    "eslint-plugin-vue": "^9.30.0",
+    "globals": "^15.11.0",
+    "postcss": "^8.4.47",
+    "sass-embedded": "^1.80.6",
+    "stylelint": "^16.10.0",
+    "tailwindcss": "^3.4.14",
+    "typescript": "~5.6.2",
+    "typescript-eslint": "^8.12.2",
+    "unplugin-auto-import": "^0.18.3",
+    "vite": "^5.4.10",
+    "vite-plugin-dts": "^4.3.0",
+    "vite-plugin-eslint2": "^5.0.1",
+    "vite-plugin-stylelint": "6.0.0-beta.1",
+    "vue-tsc": "^2.1.8"
+  }
 }
