@@ -1,74 +1,157 @@
 <template>
-  <div role="application" aria-label="Chrome color picker" :class="['vc-chrome', disableAlpha ? 'vc-chrome__disable-alpha' : '']">
+  <div
+    role="application"
+    aria-label="Chrome color picker"
+    :class="['vc-chrome', disableAlpha ? 'vc-chrome__disable-alpha' : '']"
+    :style="{ 
+      '--vc-alpha-gradient': alphaGradient,
+      '--vc-hue-gradient-horizontal': hueGradient.horizontal,
+    }"
+  >
     <div class="vc-chrome-saturation-wrap">
-      <saturation v-model="colors" @change="childChange"></saturation>
+      <!-- <PercentageSlider v-model="saturation" /> -->
     </div>
     <div class="vc-chrome-body">
       <div class="vc-chrome-controls">
         <div class="vc-chrome-color-wrap">
-          <div :aria-label="`current color is ${colors.hex}`" class="vc-chrome-active-color" :style="{background: activeColor}"></div>
-          <checkboard v-if="!disableAlpha"></checkboard>
+          <div
+            :aria-label="`current color is #${hex}`"
+            class="vc-chrome-active-color"
+            :style="{ background: `${getColorAs('hsl')}` }"
+          />
+          <Checkboard v-if="!disableAlpha" />
         </div>
 
         <div class="vc-chrome-sliders">
           <div class="vc-chrome-hue-wrap">
-            <hue v-model="colors" @change="childChange"></hue>
+            <PercentageSlider
+              v-model:percentage="hue"
+              :in-range="360"
+            />
           </div>
-          <div class="vc-chrome-alpha-wrap" v-if="!disableAlpha">
-            <alpha v-model="colors" @change="childChange"></alpha>
+          <div
+            class="vc-chrome-alpha-wrap"
+            v-if="!disableAlpha"
+          >
+            <PercentageSlider
+              v-model:percentage="alpha"
+              :in-range="1"
+            />
           </div>
         </div>
       </div>
 
-      <div class="vc-chrome-fields-wrap" v-if="!disableFields">
-        <div class="vc-chrome-fields" v-show="fieldsIndex === 0">
+      <div
+        class="vc-chrome-fields-wrap"
+        v-if="!disableFields"
+      >
+        <div
+          class="vc-chrome-fields"
+          v-show="fieldsIndex === 0"
+        >
           <!-- hex -->
           <div class="vc-chrome-field">
-            <ed-in v-if="!hasAlpha" label="hex" :modelValue="colors.hex" @change="inputChange"></ed-in>
-            <ed-in v-if="hasAlpha" label="hex" :modelValue="colors.hex8" @change="inputChange"></ed-in>
+            <EditableInput
+              v-if="alpha < 1"
+              type="hex8"
+              label="Hex"
+              v-model="hex8"
+            />
+            <EditableInput
+              v-else
+              type="hex"
+              label="Hex"
+              v-model="hex"
+            />
           </div>
         </div>
-        <div class="vc-chrome-fields" v-show="fieldsIndex === 1">
+        <div
+          class="vc-chrome-fields"
+          v-show="fieldsIndex === 1"
+        >
           <!-- rgba -->
           <div class="vc-chrome-field">
-            <ed-in label="r" :modelValue="colors.rgba.r" @change="inputChange"></ed-in>
+            <EditableInput
+              label="r"
+              v-model="red"
+            />
           </div>
           <div class="vc-chrome-field">
-            <ed-in label="g" :modelValue="colors.rgba.g" @change="inputChange"></ed-in>
+            <EditableInput
+              label="g"
+              v-model="green"
+            />
           </div>
           <div class="vc-chrome-field">
-            <ed-in label="b" :modelValue="colors.rgba.b" @change="inputChange"></ed-in>
+            <EditableInput
+              label="b"
+              v-model="blue"
+            />
           </div>
-          <div class="vc-chrome-field" v-if="!disableAlpha">
-            <ed-in label="a" :modelValue="colors.a" :arrow-offset="0.01" :max="1" @change="inputChange"></ed-in>
+          <div
+            class="vc-chrome-field"
+            v-if="!disableAlpha"
+          >
+            <EditableInput
+              label="a"
+              v-model="alpha"
+              :in-range="1"
+            />
           </div>
         </div>
-        <div class="vc-chrome-fields" v-show="fieldsIndex === 2">
+        <div
+          class="vc-chrome-fields"
+          v-show="fieldsIndex === 2"
+        >
           <!-- hsla -->
           <div class="vc-chrome-field">
-            <ed-in label="h" :modelValue="hsl.h" @change="inputChange"></ed-in>
+            <EditableInput
+              label="h"
+              type="hue"
+              v-model="hue"
+            />
           </div>
           <div class="vc-chrome-field">
-            <ed-in label="s" :modelValue="hsl.s" @change="inputChange"></ed-in>
+            <EditableInput
+              label="s"
+              type="hsla"
+              v-model="saturation"
+            />
           </div>
           <div class="vc-chrome-field">
-            <ed-in label="l" :modelValue="hsl.l" @change="inputChange"></ed-in>
+            <EditableInput
+              label="l"
+              type="hsla"
+              v-model="lightness"
+            />
           </div>
-          <div class="vc-chrome-field" v-if="!disableAlpha">
-            <ed-in label="a" :modelValue="colors.a" :arrow-offset="0.01" :max="1" @change="inputChange"></ed-in>
+          <div
+            class="vc-chrome-field"
+            v-if="!disableAlpha"
+          >
+            <EditableInput
+              label="a"
+              v-model="alpha"
+              :max="1"
+            />
           </div>
         </div>
         <!-- btn -->
-        <div class="vc-chrome-toggle-btn" role="button" aria-label="Change another color definition" @click="toggleViews">
-          <div class="vc-chrome-toggle-icon">
-            <svg style="width:24px; height:24px" viewBox="0 0 24 24"
-              @mouseover="showHighlight"
-              @mouseenter="showHighlight"
-              @mouseout="hideHighlight">
-              <path fill="#333" d="M12,18.17L8.83,15L7.42,16.41L12,21L16.59,16.41L15.17,15M12,5.83L15.17,9L16.58,7.59L12,3L7.41,7.59L8.83,9L12,5.83Z" />
-            </svg>
-          </div>
-          <div class="vc-chrome-toggle-icon-highlight" v-show="highlight"></div>
+        <div
+          class="vc-chrome-toggle-btn"
+          role="button"
+          aria-label="Change another color definition"
+          @click="incrementView"
+        >
+          +
+        </div>
+        <div
+          class="vc-chrome-toggle-btn"
+          role="button"
+          aria-label="Change another color definition"
+          @click="decrementView"
+        >
+          -
         </div>
         <!-- btn -->
       </div>
@@ -76,116 +159,158 @@
   </div>
 </template>
 
-<script>
-import colorMixin from '../mixin/color'
-import editableInput from './common/EditableInput.vue'
-import saturation from './common/Saturation.vue'
-import hue from './common/Hue.vue'
-import alpha from './common/Alpha.vue'
-import checkboard from './common/Checkboard.vue'
+<script setup lang="ts">
+import { ColorFormat } from '../composables/useColor';
+import Checkboard from './common/Checkboard.vue';
+import EditableInput from './common/EditableInput.vue';
+import PercentageSlider from './common/PercentageSlider.vue';
 
-export default {
-  name: 'Chrome',
-  mixins: [colorMixin],
-  props: {
-    disableAlpha: {
-      type: Boolean,
-      default: false
-    },
-    disableFields: {
-      type: Boolean,
-      default: false
-    }
-  },
-  components: {
-    saturation,
-    hue,
-    alpha,
-    'ed-in': editableInput,
-    checkboard
-  },
-  data () {
-    return {
-      fieldsIndex: 0,
-      highlight: false
-    }
-  },
-  computed: {
-    hsl () {
-      const { h, s, l } = this.colors.hsl
-      return {
-        h: h.toFixed(),
-        s: `${(s * 100).toFixed()}%`,
-        l: `${(l * 100).toFixed()}%`
-      }
-    },
-    activeColor () {
-      const rgba = this.colors.rgba
-      return 'rgba(' + [rgba.r, rgba.g, rgba.b, rgba.a].join(',') + ')'
-    },
-    hasAlpha () {
-      return this.colors.a < 1
-    }
-  },
-  methods: {
-    childChange (data) {
-      this.colorChange(data)
-    },
-    inputChange (data) {
-      if (!data) {
-        return
-      }
-      if (data.hex) {
-        this.isValidHex(data.hex) && this.colorChange({
-          hex: data.hex,
-          source: 'hex'
-        })
-      } else if (data.r || data.g || data.b || data.a) {
-        this.colorChange({
-          r: data.r || this.colors.rgba.r,
-          g: data.g || this.colors.rgba.g,
-          b: data.b || this.colors.rgba.b,
-          a: data.a || this.colors.rgba.a,
-          source: 'rgba'
-        })
-      } else if (data.h || data.s || data.l) {
-        const s = data.s ? (data.s.replace('%', '') / 100) : this.colors.hsl.s
-        const l = data.l ? (data.l.replace('%', '') / 100) : this.colors.hsl.l
-
-        this.colorChange({
-          h: data.h || this.colors.hsl.h,
-          s,
-          l,
-          source: 'hsl'
-        })
-      }
-    },
-    toggleViews () {
-      if (this.fieldsIndex >= 2) {
-        this.fieldsIndex = 0
-        return
-      }
-      this.fieldsIndex++
-    },
-    showHighlight () {
-      this.highlight = true
-    },
-    hideHighlight () {
-      this.highlight = false
-    }
+const emit = defineEmits(["update:color"]);
+const options = withDefaults(
+  defineProps<{
+    color: ColorFormat;
+    disableAlpha?: boolean;
+    disableFields?: boolean;
+  }>(),
+  {
+    disableAlpha: false,
+    disableFields: false,
   }
-}
+);
+
+const { color } = toRefs(options);
+const {
+  hex,
+  red,
+  green,
+  blue,
+  hue,
+  hex8,
+  saturation,
+  lightness,
+  alpha,
+  alphaGradient,
+  hueGradient,
+  getColorAs,
+} = useColor(
+  color,
+  (nc) => emit("update:color", nc)
+);
+
+
+// ["hex", "rgba", "hsla"];
+const fieldsIndex = ref(0);
+
+const incrementView = () => {
+  fieldsIndex.value = fieldsIndex.value >= 2 ? 0 : fieldsIndex.value + 1;
+};
+
+const decrementView = () => {
+  fieldsIndex.value = fieldsIndex.value < 1 ? 2 : fieldsIndex.value - 1;
+};
+
+// export default {
+//   name: 'Chrome',
+//   mixins: [colorMixin],
+//   props: {
+//     disableAlpha: {
+//       type: Boolean,
+//       default: false
+//     },
+//     disableFields: {
+//       type: Boolean,
+//       default: false
+//     }
+//   },
+//   components: {
+//     saturation,
+//     hue,
+//     alpha,
+//     'EditableInput': editableInput,
+//     checkboard
+//   },
+//   data () {
+//     return {
+//       fieldsIndex: 0,
+//       highlight: false
+//     };
+//   },
+//   computed: {
+//     hsl () {
+//       const { h, s, l } = this.colors.hsl;
+//       return {
+//         h: h.toFixed(),
+//         s: `${(s * 100).toFixed()}%`,
+//         l: `${(l * 100).toFixed()}%`
+//       };
+//     },
+//     activeColor () {
+//       const rgba = this.colors.rgba;
+//       return 'rgba(' + [rgba.r, rgba.g, rgba.b, rgba.a].join(',') + ')';
+//     },
+//     hasAlpha () {
+//       return this.colors.a < 1;
+//     }
+//   },
+//   methods: {
+//     childChange (data) {
+//       this.colorChange(data);
+//     },
+//     inputChange (data) {
+//       if (!data) {
+//         return;
+//       }
+//       if (data.hex) {
+//         this.isValidHex(data.hex) && this.colorChange({
+//           hex: data.hex,
+//           source: 'hex'
+//         });
+//       } else if (data.r || data.g || data.b || data.a) {
+//         this.colorChange({
+//           r: data.r || this.colors.rgba.r,
+//           g: data.g || this.colors.rgba.g,
+//           b: data.b || this.colors.rgba.b,
+//           a: data.a || this.colors.rgba.a,
+//           source: 'rgba'
+//         });
+//       } else if (data.h || data.s || data.l) {
+//         const s = data.s ? (data.s.replace('%', '') / 100) : this.colors.hsl.s;
+//         const l = data.l ? (data.l.replace('%', '') / 100) : this.colors.hsl.l;
+
+//         this.colorChange({
+//           h: data.h || this.colors.hsl.h,
+//           s,
+//           l,
+//           source: 'hsl'
+//         });
+//       }
+//     },
+//     toggleViews () {
+//       if (this.fieldsIndex >= 2) {
+//         this.fieldsIndex = 0;
+//         return;
+//       }
+//       this.fieldsIndex++;
+//     },
+//     showHighlight () {
+//       this.highlight = true;
+//     },
+//     hideHighlight () {
+//       this.highlight = false;
+//     }
+//   }
+// };
 </script>
 
 <style lang="css">
 .vc-chrome {
-  background: #fff;
+  /* background: #fff; */
   border-radius: 2px;
   box-shadow: 0 0 2px rgba(0,0,0,.3), 0 4px 8px rgba(0,0,0,.3);
   box-sizing: initial;
   width: 225px;
   font-family: Menlo;
-  background-color: #fff;
+  /* background-color: #fff; */
 }
 .vc-chrome-controls {
   display: flex;
@@ -240,7 +365,7 @@ export default {
   position: absolute;
   width: 24px;
   height: 28px;
-  background: #eee;
+  /* background: #eee; */
   border-radius: 4px;
   top: 10px;
   left: 12px;
@@ -249,10 +374,12 @@ export default {
   position: relative;
   height: 10px;
   margin-bottom: 8px;
+  background: var(--vc-hue-gradient-horizontal, white);
 }
 .vc-chrome-alpha-wrap {
   position: relative;
   height: 10px;
+  background: var(--vc-alpha-gradient, white);
 }
 .vc-chrome-hue-wrap .vc-hue {
   border-radius: 2px;
@@ -270,7 +397,7 @@ export default {
 }
 .vc-chrome-body {
   padding: 16px 16px 12px;
-  background-color: #fff;
+  /* background-color: #fff; */
 }
 .vc-chrome-saturation-wrap {
   width: 100%;
