@@ -1,34 +1,26 @@
 <template>
-  <div class="vc-checkerboard" :style="bgStyle"></div>
+  <div class="vc-checkerboard" :style="{ backgroundImage: `url(${checkboard})` }"></div>
 </template>
 
-<script>
-const _checkboardCache = {}
+<script setup lang="ts">
+const _checkboardCache: Record<string, string | null> = {};
 
-export default {
-  name: 'Checkboard',
-  props: {
-    size: {
-      type: [Number, String],
-      default: 8
-    },
-    white: {
-      type: String,
-      default: '#fff'
-    },
-    grey: {
-      type: String,
-      default: '#e6e6e6'
-    }
-  },
-  computed: {
-    bgStyle () {
-      return {
-        'background-image': 'url(' + getCheckboard(this.white, this.grey, this.size) + ')'
-      }
-    }
+const options = withDefaults(
+  defineProps<{
+    size?: number;
+    white?: string;
+    grey?: string;
+  }>(),
+  {
+    size: 8,
+    white: "#fff",
+    grey: "#e6e6e6",
   }
-}
+)
+
+const { size, grey, white } = toRefs(options);
+
+const checkboard = computed(() => getCheckboard(white.value, grey.value, size.value));
 
 /**
  * get base 64 data by canvas
@@ -38,18 +30,18 @@ export default {
  * @param {Number} size
  */
 
-function renderCheckboard (c1, c2, size) {
+function renderCheckboard (c1: string, c2: string, size: number) {
   // Dont Render On Server
-  if (typeof document === 'undefined') {
+  if (typeof document === 'undefined' || !globalThis.window) {
     return null
   }
+
   const canvas = document.createElement('canvas')
   canvas.width = canvas.height = size * 2
   const ctx = canvas.getContext('2d')
   // If no context can be found, return early.
-  if (!ctx) {
-    return null
-  }
+  if (!ctx) return null;
+  
   ctx.fillStyle = c1
   ctx.fillRect(0, 0, canvas.width, canvas.height)
   ctx.fillStyle = c2
@@ -67,16 +59,13 @@ function renderCheckboard (c1, c2, size) {
  * @param {Number} size
  */
 
-function getCheckboard (c1, c2, size) {
+function getCheckboard (c1: string, c2: string, size: number) {
   const key = c1 + ',' + c2 + ',' + size
 
-  if (_checkboardCache[key]) {
-    return _checkboardCache[key]
-  } else {
-    const checkboard = renderCheckboard(c1, c2, size)
-    _checkboardCache[key] = checkboard
-    return checkboard
-  }
+  if (_checkboardCache[key]) return _checkboardCache[key]
+  const checkboard = renderCheckboard(c1, c2, size)
+  _checkboardCache[key] = checkboard
+  return checkboard
 }
 
 </script>
@@ -84,10 +73,10 @@ function getCheckboard (c1, c2, size) {
 <style lang="css">
 .vc-checkerboard {
   position: absolute;
+  background-size: contain;
   top: 0px;
   right: 0px;
   bottom: 0px;
   left: 0px;
-  background-size: contain;
 }
 </style>
